Do not build TIFF when codec is disabled
Currently all of the TIFF related code is being built when support for
the codec is disabled, it just isn't being utilized. Depending on the
settings being used, this unneeded code may or may not get stripped
during linking.

This CL explicitly turns off building the TIFF codec code if support
for TIFF is turned off. It also fixes cases in the code base where tif
was being used instead of tiff.

BUG=pdfium:1080

Change-Id: If6aaa8af5160fdd5b261e63bab7d5984196efcc9