Do not build PNG when codec is disabled
Currently all of the PNG related code is being built when support for
the codec is disabled, it just isn't being utilized. Depending on the
settings being used, this unneeded code may or may not get stripped
during linking.

This CL explicitly turns off building the PNG codec code if support
for PNG is turned off.

BUG=pdfium:1080

Change-Id: I9c5247145fcadbcb1bd2243aa83350304ba421ff