Change formcalc javascript method output
This CL changes the formcalc JS conversion code so that arrays no longer
generate method code. Previously, if an object was a array we'd execute
the given method on each element and then return the value from the
first element. I can't find anywhere where this is specified. This CL
changes the code to return null if the provided object is an array.

For the referenced bug, the generated JS originally needed a buffer of
365meg to generate. With this CL, it needs a buffer of 7.5k.

Bug: chromium:814840
Change-Id: Ibb5993fa52b7c13b20b325cf8848a306f82ae014