Do not build GIF when codec is disabled
Currently all of the GIF related code is being built when support for
the codec is disabled, it just isn't being utilized. Depending on the
settings being used, this unneeded code may or may not get stripped
during linking.

This CL explicitly turns off building the GIF codec code if support
for GIF is turned off.

This also catches a few missed cases from previous CLs.

BUG=pdfium:1080

Change-Id: Ie7fe2d894d2ae2f8f36ae05e0ff256f2ce6ef8d4