Do not build BMP when codec is disabled
Currently all of the BMP related code is being built when support for
the codec is disabled, it just isn't being utilized. Depending on the
settings being used, this unneeded code may or may not get stripped
during linking.

This CL explicitly turns off building the BMP codec code if support
for BMP is turned off.

BUG=pdfium:1080

Change-Id: I56d40639a5a3631f9c601a1eef3f98873feac94f