Cleanup formcalc method generation
This CL simplifies the FormCalc method call generation when converted to
JavaScript. Currently we'd output the same chunk of code to run the
given method on an array or object per method call. This CL pulls out
the common execution code to a pfm_method_runner function which is used
instead.

An embedder test has been added to verify that method invocation from
fromcalc works correctly.

Bug: chromium:814848
Change-Id: I1ec052eab051053fedcb464d57e0e15228b8c5a2