Use internal wcstof instead of system wcstod in formcalc lexer
This CL switches the usage of wcstod to use the FXSYS_wcstof to
determine if a given string is a valid floating point number.

Using the internal method makes linux slightly slower (10's of ms)
makes mac a lot faster 900ms to 60ms for the test case in the bug.

Bug: chromium:813646
Change-Id: Ie68287a336e3b95a0c0b845d5bf39db6fc82b39c