Add support for PartionRealloc to return nullptr
Currently the PartitionRealloc code path will only OOM, with no option
to return nullptr on failure, unlike PartitionAlloc code path. This CL
refactors the realloc code path to be similar to alloc code path,
following the upstream patch:
https://chromium-review.googlesource.com/c/chromium/src/+/1044971

This also changes the version of realloc exposed to third party C libs
to have the nullptr behaviour, like the exposed version of alloc.

This CL is redo of
https://pdfium-review.googlesource.com/c/pdfium/+/31990

BUG=chromium:783022

Change-Id: Ib1b659079585dfd0423d683b8a2c7b6758a22a01