Make CPDF_Page retainable.
Small step to reducing the differences between XFA and non-XFA. We
still use the RetainPtr pretty much as if it were an unique_ptr, in
that we're not yet caching pages and handing out multiple pointers
to the same page in the non-XFA case.

The one change is in page view cleanup, where we no longer need a
boolean and can take (sufficient) page ownership with a RetainPtr.

NOTE: Wait for imminent branch to pass before landing.

Change-Id: I64a2f12ac3424ece1063d40583995b834117cf34