Make CPDF_Document cache CPDF_Pages
We cache pages not by page number, which can bounce around as pages
are inserted or removed, but by page dictionary's object number.

Since the page may be created under one function and used under
another, we can't take the shortcut of not instantiating a render
cache nor not parsing the page.

Change-Id: I9a325cda8b3141153544ac53e78a51a44e6b411a