[xfa] Allow accessing the event changed parameter from JS.
This CL updates the CFXJSE_Engine code to access the defined properties
of a class if we fail to find the property in any other way. This fixes
up an issue where we were unable to read the 'change' property of the
CJX_EventPseudoModel because we could not find the 'change' property.

Bug: 1066
Change-Id: I4ad205bc527beeca1c3e24a36cdde0c21287d9fb