Reserve space in code words when creating encoder context
During encoding, the code is going to walk through every character in
m_msg and determine what values need to be added to m_codewords. The
ceiling on the number of characters being added to m_codewords is the
length of m_msg, so reserving enought space in advance. This prevents
thrash related to incrementing the string and thus resizing in a tight
loop.

BUG=chromium:846027

Change-Id: Icefd6955933f8068feeeee6243e430f60c50d747