Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Do you want the upstream folks to take a look, especially since you are copying the OWNERS file with their names in it?

I am really not sure what out policy in this regards is. Unfortunately the files are not in a separate repo that I can DEPS in, so I have just copied the files accross and made no changes to them. I was following what appeared to be done in tools/clang, which has the OWNERS file in it. Should I have not brought across the OWNERS file?

mmoroz@, who is CC'd on this CL, I think would be an appropriate person to review if we want an upstream reviewer.

Patch-set: 1
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0