Implement handling --font-dir on Win
Currently the value passed in via the flag is ignored on Win. This
implements the needed logic to use the given path instead of the
standard system font directories.

There still exists a bug on Win, where XFA is still doing font
substitution, even when the font is present in the font
directory. Putting tracing in shows that the fonts are being found and
loaded correctly, and it is a bug somewhere else with how font
finding/substituion is performed.

BUG=pdfium:1095

Change-Id: I1a817e67516a93e8df2bcd70cca5f85f97938a5a