Migrate coverage_report.py to use upstream Chromium scripts
This cherry picks in the Chromium coverage scripts and converts our
existing coverage_report.py to use them instead of gcov & lcov. This
generates a different format of HTML report, but the content appears
to be the same. Some of the coverage numbers changed a bit, due to
differences in how llvm-cov and gcov calculate executable lines, but
drilling down into the reports I think llvm-cov is more accurate
overall and there are no major descrepencies.

The upstream code has been added to tools/, because that appears to be
where other scripts we have pulled in are living, and it means no
patching is needed to the upstream code.

Large portions of the existing script are left as is and just the
report generation has been changed. I plan in follow up CLs to remove
the duplication of functionality in the PDFium scripts and modularlize
the upstream code better.

BUG=pdfium:1069

Change-Id: I009bfb8aac8f1a878e01ff70923e19bbb4774a9c