GCC6: workaround bug with external class template linking.
In the case of ByteStringView and WideStringView, non official
build fails to link in GCC 6.3 or older. But it works in GCC 7.2.

A workaround for this problem is avoiding declaring in same file
extern and explicit class template for them.

It seems to be related to https://gcc.gnu.org/bugzilla/show_bug.cgi?id=57728
though that bug is supposed to be already fixed in 6.3.0.

Change-Id: Ia6323f23138c0c11e59ce7aa31da1f6c11bc4a5a