Update patch set 1
Patch Set 1:

My main question here is, we don't know why this is here in the first place, so do we have any idea of what this might break? Was it added because existing documents have this behaviour? What's the chances we break something be doing this?

Patch-set: 1
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>