Migrate coverage_report.py to use upstream Chromium scripts
This adds tools/code_coverage from Chromium to DEPSand converts our
existing coverage_report.py to use it instead of gcov & lcov. This
generates a different format of HTML report, but the content appears
to be the same. Some of the coverage numbers changed a bit, due to
differences in how llvm-cov and gcov calculate executable lines, but
drilling down into the reports I think llvm-cov is more accurate
overall and there are no major descrepencies.

Large portions of the existing script are left as is and just the
report generation has been changed. I plan in follow up CLs to remove
the duplication of functionality in the PDFium scripts and modularlize
the upstream code better.

BUG=pdfium:1069

Change-Id: I009bfb8aac8f1a878e01ff70923e19bbb4774a9c