Rename CPDFXFA_Context::GetTimerMgr() to NewTimerMgr()
Because it returns an unique_ptr to a newly-created object.

Also rename m_pTimeMgrAdapter to m_pAdapterTimerMgr, because in
English, adjectives come first, and this is a TimerMgr obtained
from an Adatper, not an Adapter for a TimerMgr. Also change TimeMgr
to TimerMgr since the manager manages timers, not time itself.

Follow-on from https://pdfium-review.googlesource.com/c/pdfium/+/33230
No functional change.

Change-Id: I08d2b6cd8cc816f38326e87ee46060f091c74fdf