Fix a number of unused writes in barcode code
This fixes a number of instances where a value is written out to a
variable, but never used in the barcode. There are three different
types of fixes employed in this CL. If the non-use is a bug, then
rewrite the code to actually use the value. If it is an assignment
with no side effects, then just remove the entire line. Finally if it
is an assignment of a function return value, cast it to void instead
to clearly mark that the return is being ignored.

Change-Id: If13a3684cb2db81592cce9a798788a26fcdb4c6d