Add ability to log click events reaching PDFium
Adds in logging statement to related API methods and a build flag to
control enabling the feature. This is meant to be used in conjunction
with Chromium.

A developer can enable this flag, load a test PDF, and then click on
elements to determine appropriate coordinates for an .evt file.

Depends on
https://chromium-review.googlesource.com/c/chromium/src/+/1087394
being landed.

BUG=pdfium:1057

Change-Id: I956c39f545297872a61affdb621a412b3c407ff5