Update patch set 4
Patch Set 4:

This as written seems like the simplest way to do this. We don't really need to pipe this out to the embedder as it isn't something most folks would want to do. Same with making it runtime, we probably don't ever want this in released chrome, so having a build flag is the simplest solution.

If we find we nee these in other places we can look at getting more complicated but keeping it simple for now seems like the easiest path forward.

Patch-set: 4
Reviewer: Gerrit User 5965 <5965@fca2d64a-0420-3532-b200-12848c95a47b>