Update patch set 2
Patch Set 2:

core/fpdfapi/parser/cpdf_hint_tables.cpp:64 removes a check that nStreamOffset is non-negative, FX_FILESIZE is a signed value.

core/fpdfapi/parser/cpdf_hint_tables.cpp:168 removes a check that nOffsetE is non-negative.
There should also be a check around here that nFirstPageNum is in the range defined by nPages.

core/fpdfapi/parser/cpdf_hint_tables.cpp:241 removes a check that nStreamOffset is non-negative and a check that nStreamLen is positive and non-zero.

Adding in the check that nFirstPageNum is in the range defined by nPages is what I originally did in https://pdfium-review.googlesource.com/c/pdfium/+/34310, but on further reflection, I am not confident that removing these checks isn't going to cause other issues, so I decided it was more prudent to revert the entire CL.

Patch-set: 2