Update patch set 1
Patch Set 1:

Not sure if this is exactly right, but it does all compile and link. Note that Skia is now depending on c++14 language features, so it may be necessary to comment out the 'use_cxx11 = true' line in the .gn file to build in addition to this. I started on this just to move SkFontMgr.h from ports/ to core/, since we need that to land before Skia can remove the old ports/SkFontMgr.h forwarding header. Feel free to take this and fix it up if needed.

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Reviewer: Gerrit User 5845 <5845@fca2d64a-0420-3532-b200-12848c95a47b>