Update patch set 4
Patch Set 4: Commit-Queue+2

> Patch Set 3: Code-Review+1
> 
> > Patch Set 3:
> > 
> > lg but can we get some kind of test?
> 
> Isn't this covered by 
> https://cs.chromium.org/chromium/src/third_party/pdfium/fxjs/cfxjse_formcalc_context_embeddertest.cpp?rcl=d9fb538bc08c4923fbfd3d50c47a61716d679956&l=1158

This specific case wasn't covered. 
The root of the issue is that you can specify some characters like Å two different ways, either as a code point for that character or as two, one for A and one indicating the A have a o over it. When converting to UTF8 in Lower, the second case is being covnerted into the first case. This means the converted string is shorter then the original, and we were using the original length for iteration. I have added a test for Lower("ÅÅÅ") which gets caught by ASAN on HEAD but is fixed by this CL.

Patch-set: 4
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=+2