Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> (2 comments)
> 
> > Patch Set 5: Code-Review+1
> > 
> > I worry that maybe we'll tie ourselves into knots jumping into the UnSup handler while JS is still running, but we'll find out.
> 
> I am going to implement the Chrome side of this on Monday, so hopefully I will know soon if this solution is problematic. I will not submit CL until I am confident the Chrome implementation will work.

I have taken a bit of a look at the Chromium side of this. Using the existing UnSupported Handler as is I don't think is the correct way to go. The existing handler is meant to be used for when loading the document and reporting unsupported features. It uses a global that points to the PDFium engine on the Chromium side that is set right before the load and unset right after it completes. Thus it is not set when the PDF is interactive. Because of the global nature of this value there is not a good way for PDFium to direct the call to the correct engine, it is the engine telling PDFium to send to it.

My thinking was that the correct way to do this would be to plumb through the call in PDFium and have the Chromium side see a call to save and throw up the warning. This would resolve the current issue with translation and if done correctly would leave a path to having this warning for non-XFA and non-Menu calls to SaveAs.

Looking at the existing code there is no such API at all and I am honestly not sure on what it should look like. It feels like I am starting to go down the path of just implementing SaveAs, which is not in scope for XFA launch. Should I just implement an API method like SaveAsCalled, or something like that, with documentation indicating it should be removed when the feature is actually implemented?

Patch-set: 6
Reviewer: Gerrit User 7065 <7065@fca2d64a-0420-3532-b200-12848c95a47b>