Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> Lei, PTAL.
> Art, FYI, this was one of your to-dos that we can remove.

The problem was not in the storing of CPDF_IndirectObjectHolder within CPDF_ObjectAvail. This is just the consequence of the original problem. The original problem is in PDFium API. The problem Is - We should destroy the FPDF_AVAIL after FPDF_DOCUMENT. And in this case the FPDF_DOCUMENT is CPDF_IndirectObjectHolder and CPDF_ObjectAvail is storing within CPDF_DataAvail (which is FPDF_AVAIL).

I think, that this solution is increased the complexity of current code(we should pass additional param), and it is not bringed us to close original problem.

Patch-set: 2
Reviewer: Gerrit User 7160 <7160@fca2d64a-0420-3532-b200-12848c95a47b>