Fix hint tables parsing.
Sample PDF:
https://yadi.sk/d/oWLtAEfy3YbEb3

For offsets, equal to the hint stream offset, added hint stream
length to determine the actual offset, because the hint stream
may be stored on the same place, where was objcect in
moment, when hints tables calculated.

Also the number of bits needed to represent the numerator of
the fractional position for each shared object reference may be
zero, if each shared group contains only one object with obj num,
incremented on 1.

Change-Id: I4754d603f388354821e8d0cac97ad99a7578fe4b