Add pdfium::base::Vector2D<>()
Many of the FX_Alloc's that have not been converted to std::vector
are using FX_Alloc2D and the safe math it performs under the covers.
Make an equivalent function for returning a vector to avoid burdening
callers with the safe math equivalents.

Use it in one file for PoC.

Change-Id: Ie4a280351b7370b755f2a1928f8b2d84fe007c03