Make MarkData::mMarks an std::vector<RetainPtr<CPDF_ContentMarkItem>>
Copying MarkData whenever a content mark is opened or closed is
inneficient since each MarkData will have copies of the same
CPDF_ContentMarkItems.

This CL changes the vector inside MarkData to have only the pointers
to the CPDF_ContentMarkItems rather than copies of it. This also
unifies the dictionaries of each content mark. Previously, there were
as many copies of it as content marks scope changes inside it.

Bug: pdfium:1037
Change-Id: Ia6f79b401d4e14ac07dbba81bbd97df965b77c94