Clean up ProcessFormatTestValidate
The existing implementation was overly complex and hard to understand,
so this simplifies the logic. As part of this a number of issues, like
there being no failure path, have been resolved.

The spec implies that a picture clause of "" should only accept the
empty string, but existing implementation returns that validation does
exist. This is due to the GetPicture() method returning a string not a
pointer or Optional, so there is no mechanism to differentiate between
a clause with an empty string and the clause not being present. This
CL maintains the existing behaviour, because there may be code else
where that depends on it.

The existing implementation returns validation not existing if the
node is marked as non-interactable. Though this seems intuitively
correct, it is problematic, because this logic isn't just called via
an interaction with the node. The validitity check could be initiated
by a call from JS attached to an event or different node. Thus it is
possible for a node under test to be non-interactive, but the result
of the test is still important. The caller may make the node
interactive if it isn't valid for example. The spec doesn't state
anything about interactivity controlling if validity checks run. This
CL removes the related logic, since it was causing issues with an
example PDF from the wild.

BUG=pdfium:1065

Change-Id: Icb5b97e0d90f6fbc7ad6b87d81e256803c757eb0