Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I created the pdf myself. First I typed a hello world PDF 1.2 in vim, then used pdftk to insert the byte offsets, finally used pdfcrop from tex to create the form object.

For manually created PDFs that are used in embeddertests, we normally use a .in file and then call testing/tools/fixup_pdf_template.py to create the PDF file. Do you mind doing this? There are some examples of .in files in the testing/resources folder. If it becomes too much of a hassle then we can just land as-is (we already have some files that are only PDF and don't have a .in).

Patch-set: 1
Reviewer: Gerrit User 6295 <6295@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0