Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > I recently noticed Chromium's ui/ code use this technique, and one of the modern/effective C++ books also mention it. WDYT?
> 
> I thought we were on a crusade to get rid of const_cast? Should we open an exception for this pattern? It's not exactly safe.

Yes, in general I'm not a fan of const_cast, but I see the convenience of doing this for duplicate const/non-const functions. So I'm tossing it out for discussion.

Why is it not safe?

Patch-set: 1
Reviewer: Gerrit User 5045 <5045@fca2d64a-0420-3532-b200-12848c95a47b>
Label: Commit-Queue=0