Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I recently noticed Chromium's ui/ code use this technique, and one of the modern/effective C++ books also mention it. WDYT?
> > 
> > I thought we were on a crusade to get rid of const_cast? Should we open an exception for this pattern? It's not exactly safe.
> 
> Yes, in general I'm not a fan of const_cast, but I see the convenience of doing this for duplicate const/non-const functions. So I'm tossing it out for discussion.
> 
> Why is it not safe?

My understanding is:
One can run a const method that changes the instance state with this technique. That's bad by itself.

Even assuming it's all correct at first (the non-const version could be const), if the non-const version is changed and the new version changes the instance state, the assumption that state wouldn't change is broken for the const version. It's easy to miss there is a const version when making this change.

Patch-set: 1
Reviewer: Gerrit User 7035 <7035@fca2d64a-0420-3532-b200-12848c95a47b>