Update patch set 27
Patch Set 27:

> Patch Set 27:
> 
> > Patch Set 27:
> > Have you tried this with the latest Chromium code? It doesn't actually render the annotations for me on Windows or Linux. Maybe something changed in the Chromium font code?
> 
> I filed https://crbug.com/862051 with my findings. However, that bug is Linux-only and does not explain the behavior on Windows. Should I (or you) sync Chromium back a couple of months and try this CL there?

BTW: The ansi code page of my WinOS is cyrrlic. And as result the selected for ACP font is suitable for annotations within specified file.
But in case when other ansi code page is selected in WinOS, which is not suitable for annotations, the font may be selected incorrectly.

I think we should detect font not by ACP,  we should select it by processing annotation data, and detection of its language. But for now i dare to say, that we can use ACP as frequent case.

Patch-set: 27
Reviewer: Gerrit User 7160 <7160@fca2d64a-0420-3532-b200-12848c95a47b>