Reserve space for result in ::FilterContents
This changes the implementation for the specific bug listed and
proactively fixes it in the other overrides. The general bug here is
that if you concat a WideString in a tight loop without first
reserving space, each call will cause an allocation size change and
memcpy. This is very expensive and causes ClusterFuzz cases to
timeout.

BUG=chromium:863295

Change-Id: I6c1d900a31b98cd9ddcf91d1ec0f3973c9cdfa26