Replace downcasts in cxfa_node to checked conversion methods.
Part of the effort to not trust downcasts in the CFXA_Node hieracrhy.
The number of subclasses is small enough here that the O(n^2) space
overhead for the vtables is not significant.

Put all sub-classes in same top-level namespace as the parent, just
because we avoid jumping in and out of anonymous namespace.

Change-Id: I32b80a93c75f8a09eb17c01fe1e29d04b4053fa1