Update patch set 10
Patch Set 10:

With this changes, what can be now the best method to add xref table consolidation, when saving, in order to get rid of the free obj references? The patch I submitted in the discussion of this subject, [issue 1076](https://bugs.chromium.org/p/pdfium/issues/detail?id=1076#c11), made use of code, in the cpdf_creator files, that is now removed/reworked with this commit, that was really suitable for the task.

Also, can you take a look to this commit code related issue (https://bugs.chromium.org/p/pdfium/issues/detail?id=1102).

Patch-set: 10
CC: Gerrit User 7407 <7407@fca2d64a-0420-3532-b200-12848c95a47b>