Type-safe iteration over CXFA_ContentLayoutItems.
There's no reason that every CXFA_LayoutItem must be a
CXFA_ContentLayout item, so use checked conversions. The
code may already be arranging things so the constraint is
already satisfied, but this adds an additional layer of
safety.

Tidy some long expressions with pre-existing typedef for node
iterator while we're at it.

Change-Id: If1448131ca73bfedb2266b070f1a913688598e64