Bounds check srcline in Big2_Image.cpp.
No matter how the dimensions might be determined, we know the
hard end of the source line, and can use it for a bounds check.
We expect the size is quantized to a multiple of m_stride, so
as long as each block operates within an m_stride, the initial
check should be sufficient.

Bug: 867501
Change-Id: Iaf9936557b856f3eb09fef522f3e6738aa4f38f0