Limit size of expression list in FormCalc parser
Limits the number of elements that can be added to the expressions
list in the FormCalc parser. This handles cases like long strings of !
repeated, since ! is a valid identifier and identifiers are valid
expression, even though it will be no-op. This is another case of
something that is valid, but stupid.

BUG=chromium:870385

Change-Id: I8e34ce00bcbe4499e0a45bd5dc38541793144481