Add proxy for syscall time
This CL adds a proxy, FXSYS_time, for the time syscall, so that a
testing mechanism can be implemented. Specically there is now a flag
for pdfium_test, --time=, that allows setting the time since the epoch
that will be returned. This plumbed all the way down into the proxy
and allows for stable results for tests that depend on getting the
current time.

There are other places in the code base that will need to be patched
like this, that will be dealt with in follow on CLs.

BUG=pdfium:1104

Change-Id: I2de185f8d47abe46704dd579c13a54948b7f81e0