Optimize rendering of two dimensional barcodes.
Reduce the logical size of the barcode by removing unnecessary
region duplication.

As far as I can tell, the line thickness is useless and the aspect
ratio causes arbitrary changes in rounding, but ultimately the
dimensions of a barcode are defined by its width and height, rather
than by this ratio.

Also remove rotation code to simplify, though we don't have any
sample PDFs with this.

The improvement with this CL is from ~580ms to ~390ms per barcode,
so about 1.5x. Combined with
https://pdfium-review.googlesource.com/c/pdfium/+/40010
the improvement is to ~15ms, which is about 39x.

Change-Id: Ic28e60f54719552cfe69ace7ebc3f730c338a129